requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
Django==1.11.4
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
ecdsa==0.13
idna==2.6
paramiko==2.2.1
pem==17.1.0
pyasn1==0.3.2
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.2
PyNaCl==1.1.2
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
